+++ src/components/index.ts
@@ -2,3 +2,12 @@ export * from '@/components/atoms';
 export * from '@/components/molecules';
 export * from '@/components/organisms';
 export * from '@/components/skeletons';
+import Error from '@/components/error/Error';
+import ApiErrorFallback from '@/components/fallback/ApiErrorFallback';
+import GlobalErrorFallback from '@/components/fallback/GlobalErrorFallback';
+
+export {
+	Error,
+	GlobalErrorFallback,
+	ApiErrorFallback,
+};
