+++ src/utils/index.ts
@@ -1,6 +1,3 @@
-import ApiErrorFallback from '@/utils/fallback/apiErrorFallback';
-import GlobalErrorFallback from '@/utils/fallback/globalErrorFallback';
-
 export * from '@/utils/api-service/portfolio';
 export * from '@/utils/api-service/users';
 export * from '@/utils/api-service/commission';
@@ -13,4 +10,3 @@ export * from '@/utils/wheelHandler';
 export * from '@/utils/toast';
 export * from '@/utils/transform';
 export * from '@/utils/error';
-export { GlobalErrorFallback, ApiErrorFallback };
